+++ R/onekp.R
@@ -1,8 +1,12 @@
 #' @importFrom rlang .data
-#' @importFrom magrittr "%>%"
 #' @importFrom methods new
 #' @importFrom utils head untar
 utils::globalVariables(c("%>%", "."))
+NULL
+
+#' @importFrom magrittr %>%
+#' @export
+magrittr::`%>%`
 NULL
 
 #' onekp: access data from the 1KP project